rename MirroringAudioTrackLookup -> MirroringAudioTrackResolver & move some minor stuff around
