+++ main/src/main/java/com/github/topisenpai/lavasrc/deezer/DeezerAudioTrack.java
@@ -20,87 +20,87 @@ import java.security.NoSuchAlgorithmException;
 
 public class DeezerAudioTrack extends DelegatedAudioTrack {
 
-    private final String isrc;
-    private final String artworkURL;
-    private final DeezerAudioSourceManager sourceManager;
+	private final String isrc;
+	private final String artworkURL;
+	private final DeezerAudioSourceManager sourceManager;
 
-    public DeezerAudioTrack(AudioTrackInfo trackInfo, String isrc, String artworkURL, DeezerAudioSourceManager sourceManager) {
-        super(trackInfo);
-        this.isrc = isrc;
-        this.artworkURL = artworkURL;
-        this.sourceManager = sourceManager;
-    }
+	public DeezerAudioTrack(AudioTrackInfo trackInfo, String isrc, String artworkURL, DeezerAudioSourceManager sourceManager) {
+		super(trackInfo);
+		this.isrc = isrc;
+		this.artworkURL = artworkURL;
+		this.sourceManager = sourceManager;
+	}
 
-    public String getISRC() {
-        return this.isrc;
-    }
+	public String getISRC() {
+		return this.isrc;
+	}
 
-    public String getArtworkURL() {
-        return this.artworkURL;
-    }
+	public String getArtworkURL() {
+		return this.artworkURL;
+	}
 
-    private URI getTrackMediaURI() throws IOException, URISyntaxException {
-        var getSessionID = new HttpPost(DeezerAudioSourceManager.PRIVATE_API_BASE + "?method=deezer.ping&input=3&api_version=1.0&api_token=");
-        var json = HttpClientTools.fetchResponseAsJson(this.sourceManager.getHttpInterface(), getSessionID);
-        if (json.get("data").index(0).get("errors").index(0).get("code").asLong(0) != 0) {
-            throw new IllegalStateException("Failed to get session ID");
-        }
-        var sessionID = json.get("results").get("SESSION").text();
+	private URI getTrackMediaURI() throws IOException, URISyntaxException {
+		var getSessionID = new HttpPost(DeezerAudioSourceManager.PRIVATE_API_BASE + "?method=deezer.ping&input=3&api_version=1.0&api_token=");
+		var json = HttpClientTools.fetchResponseAsJson(this.sourceManager.getHttpInterface(), getSessionID);
+		if (json.get("data").index(0).get("errors").index(0).get("code").asLong(0) != 0) {
+			throw new IllegalStateException("Failed to get session ID");
+		}
+		var sessionID = json.get("results").get("SESSION").text();
 
-        var getUserToken = new HttpPost(DeezerAudioSourceManager.PRIVATE_API_BASE + "?method=deezer.getUserData&input=3&api_version=1.0&api_token=");
-        getUserToken.setHeader("Cookie", "sid=" + sessionID);
-        json = HttpClientTools.fetchResponseAsJson(this.sourceManager.getHttpInterface(), getUserToken);
-        if (json.get("data").index(0).get("errors").index(0).get("code").asLong(0) != 0) {
-            throw new IllegalStateException("Failed to get user token");
-        }
-        var userLicenseToken = json.get("results").get("USER").get("OPTIONS").get("license_token").text();
-        var apiToken = json.get("results").get("checkForm").text();
+		var getUserToken = new HttpPost(DeezerAudioSourceManager.PRIVATE_API_BASE + "?method=deezer.getUserData&input=3&api_version=1.0&api_token=");
+		getUserToken.setHeader("Cookie", "sid=" + sessionID);
+		json = HttpClientTools.fetchResponseAsJson(this.sourceManager.getHttpInterface(), getUserToken);
+		if (json.get("data").index(0).get("errors").index(0).get("code").asLong(0) != 0) {
+			throw new IllegalStateException("Failed to get user token");
+		}
+		var userLicenseToken = json.get("results").get("USER").get("OPTIONS").get("license_token").text();
+		var apiToken = json.get("results").get("checkForm").text();
 
-        var getTrackToken = new HttpPost(DeezerAudioSourceManager.PRIVATE_API_BASE + "?method=song.getData&input=3&api_version=1.0&api_token=" + apiToken);
-        getTrackToken.setEntity(new StringEntity("{\"sng_id\":\"" + this.trackInfo.identifier + "\"}", ContentType.APPLICATION_JSON));
-        json = HttpClientTools.fetchResponseAsJson(this.sourceManager.getHttpInterface(), getTrackToken);
-        if (json.get("data").index(0).get("errors").index(0).get("code").asLong(0) != 0) {
-            throw new IllegalStateException("Failed to get track token");
-        }
-        var trackToken = json.get("results").get("TRACK_TOKEN").text();
+		var getTrackToken = new HttpPost(DeezerAudioSourceManager.PRIVATE_API_BASE + "?method=song.getData&input=3&api_version=1.0&api_token=" + apiToken);
+		getTrackToken.setEntity(new StringEntity("{\"sng_id\":\"" + this.trackInfo.identifier + "\"}", ContentType.APPLICATION_JSON));
+		json = HttpClientTools.fetchResponseAsJson(this.sourceManager.getHttpInterface(), getTrackToken);
+		if (json.get("data").index(0).get("errors").index(0).get("code").asLong(0) != 0) {
+			throw new IllegalStateException("Failed to get track token");
+		}
+		var trackToken = json.get("results").get("TRACK_TOKEN").text();
 
-        var getMediaURL = new HttpPost(DeezerAudioSourceManager.MEDIA_BASE + "/get_url");
-        getMediaURL.setEntity(new StringEntity("{\"license_token\":\"" + userLicenseToken + "\",\"media\": [{\"type\": \"FULL\",\"formats\": [{\"cipher\": \"BF_CBC_STRIPE\", \"format\": \"MP3_128\"}]}],\"track_tokens\": [\"" + trackToken + "\"]}", ContentType.APPLICATION_JSON));
-        json = HttpClientTools.fetchResponseAsJson(this.sourceManager.getHttpInterface(), getMediaURL);
-        if (json.get("data").index(0).get("errors").index(0).get("code").asLong(0) != 0) {
-            throw new IllegalStateException("Failed to get media URL");
-        }
-        return new URI(json.get("data").index(0).get("media").index(0).get("sources").index(0).get("url").text());
-    }
+		var getMediaURL = new HttpPost(DeezerAudioSourceManager.MEDIA_BASE + "/get_url");
+		getMediaURL.setEntity(new StringEntity("{\"license_token\":\"" + userLicenseToken + "\",\"media\": [{\"type\": \"FULL\",\"formats\": [{\"cipher\": \"BF_CBC_STRIPE\", \"format\": \"MP3_128\"}]}],\"track_tokens\": [\"" + trackToken + "\"]}", ContentType.APPLICATION_JSON));
+		json = HttpClientTools.fetchResponseAsJson(this.sourceManager.getHttpInterface(), getMediaURL);
+		if (json.get("data").index(0).get("errors").index(0).get("code").asLong(0) != 0) {
+			throw new IllegalStateException("Failed to get media URL");
+		}
+		return new URI(json.get("data").index(0).get("media").index(0).get("sources").index(0).get("url").text());
+	}
 
-    private byte[] getTrackDecryptionKey() throws NoSuchAlgorithmException {
-        var md5 = Hex.encodeHex(MessageDigest.getInstance("MD5").digest(this.trackInfo.identifier.getBytes()), true);
-        var master_key = this.sourceManager.getMasterDecryptionKey().getBytes();
+	private byte[] getTrackDecryptionKey() throws NoSuchAlgorithmException {
+		var md5 = Hex.encodeHex(MessageDigest.getInstance("MD5").digest(this.trackInfo.identifier.getBytes()), true);
+		var master_key = this.sourceManager.getMasterDecryptionKey().getBytes();
 
-        var key = new byte[16];
-        for (int i = 0; i < 16; i++) {
-            key[i] = (byte) (md5[i] ^ md5[i + 16] ^ master_key[i]);
-        }
-        return key;
-    }
+		var key = new byte[16];
+		for (int i = 0; i < 16; i++) {
+			key[i] = (byte) (md5[i] ^ md5[i + 16] ^ master_key[i]);
+		}
+		return key;
+	}
 
-    @Override
-    public void process(LocalAudioTrackExecutor executor) throws Exception {
-        try (var httpInterface = this.sourceManager.getHttpInterface()) {
-            try (var stream = new DeezerPersistentHttpStream(httpInterface, this.getTrackMediaURI(), this.trackInfo.length, this.getTrackDecryptionKey())) {
-                processDelegate(new Mp3AudioTrack(this.trackInfo, stream), executor);
-            }
-        }
-    }
+	@Override
+	public void process(LocalAudioTrackExecutor executor) throws Exception {
+		try (var httpInterface = this.sourceManager.getHttpInterface()) {
+			try (var stream = new DeezerPersistentHttpStream(httpInterface, this.getTrackMediaURI(), this.trackInfo.length, this.getTrackDecryptionKey())) {
+				processDelegate(new Mp3AudioTrack(this.trackInfo, stream), executor);
+			}
+		}
+	}
 
-    @Override
-    protected AudioTrack makeShallowClone() {
-        return new DeezerAudioTrack(this.trackInfo, this.isrc, this.artworkURL, this.sourceManager);
-    }
+	@Override
+	protected AudioTrack makeShallowClone() {
+		return new DeezerAudioTrack(this.trackInfo, this.isrc, this.artworkURL, this.sourceManager);
+	}
 
-    @Override
-    public AudioSourceManager getSourceManager() {
-        return this.sourceManager;
-    }
+	@Override
+	public AudioSourceManager getSourceManager() {
+		return this.sourceManager;
+	}
 
 }
