+++ main/src/main/java/com/github/topisenpai/lavasrc/mirror/MirroringAudioTrackResolver.java
@@ -1,0 +1,9 @@
+package com.github.topisenpai.lavasrc.mirror;
+
+import com.sedmelluq.discord.lavaplayer.track.AudioItem;
+
+import java.util.function.Function;
+
+@FunctionalInterface
+public interface MirroringAudioTrackResolver extends Function<MirroringAudioTrack, AudioItem> {
+}
