+++ main/src/main/java/com/github/topisenpai/lavasrc/mirror/TrackNotFoundException.java
@@ -2,6 +2,6 @@ package com.github.topisenpai.lavasrc.mirror;
 
 public class TrackNotFoundException extends RuntimeException {
 
-    private static final long serialVersionUID = 6550093849278285754L;
+	private static final long serialVersionUID = 6550093849278285754L;
 
 }
