+++ main/src/main/java/com/github/topisenpai/lavasrc/spotify/SpotifySourceManager.java
@@ -1,8 +1,8 @@
 package com.github.topisenpai.lavasrc.spotify;
 
-import com.github.topisenpai.lavasrc.mirror.lookup.DefaultMirroringAudioTrackLookup;
-import com.github.topisenpai.lavasrc.mirror.lookup.MirroringAudioTrackLookup;
+import com.github.topisenpai.lavasrc.mirror.DefaultMirroringAudioTrackResolver;
 import com.github.topisenpai.lavasrc.mirror.MirroringAudioSourceManager;
+import com.github.topisenpai.lavasrc.mirror.MirroringAudioTrackResolver;
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager;
 import com.sedmelluq.discord.lavaplayer.tools.DataFormatTools;
 import com.sedmelluq.discord.lavaplayer.tools.JsonBrowser;
@@ -38,262 +38,262 @@ import java.util.stream.Collectors;
 
 public class SpotifySourceManager extends MirroringAudioSourceManager implements HttpConfigurable {
 
-    public static final Pattern URL_PATTERN = Pattern.compile("(https?://)?(www\\.)?open\\.spotify\\.com/(user/[a-zA-Z0-9-_]+/)?(?<type>track|album|playlist|artist)/(?<identifier>[a-zA-Z0-9-_]+)");
-    public static final String SEARCH_PREFIX = "spsearch:";
-    public static final String RECOMMENDATIONS_PREFIX = "sprec:";
-    public static final int PLAYLIST_MAX_PAGE_ITEMS = 100;
-    public static final int ALBUM_MAX_PAGE_ITEMS = 50;
-    public static final String API_BASE = "https://api.spotify.com/v1/";
-    private static final Logger log = LoggerFactory.getLogger(SpotifySourceManager.class);
+	public static final Pattern URL_PATTERN = Pattern.compile("(https?://)?(www\\.)?open\\.spotify\\.com/(user/[a-zA-Z0-9-_]+/)?(?<type>track|album|playlist|artist)/(?<identifier>[a-zA-Z0-9-_]+)");
+	public static final String SEARCH_PREFIX = "spsearch:";
+	public static final String RECOMMENDATIONS_PREFIX = "sprec:";
+	public static final int PLAYLIST_MAX_PAGE_ITEMS = 100;
+	public static final int ALBUM_MAX_PAGE_ITEMS = 50;
+	public static final String API_BASE = "https://api.spotify.com/v1/";
+	private static final Logger log = LoggerFactory.getLogger(SpotifySourceManager.class);
 
-    private final HttpInterfaceManager httpInterfaceManager = HttpClientTools.createDefaultThreadLocalManager();
-    private final String clientId;
-    private final String clientSecret;
-    private final String countryCode;
-    private String token;
-    private Instant tokenExpire;
+	private final HttpInterfaceManager httpInterfaceManager = HttpClientTools.createDefaultThreadLocalManager();
+	private final String clientId;
+	private final String clientSecret;
+	private final String countryCode;
+	private String token;
+	private Instant tokenExpire;
 
-    public SpotifySourceManager(String[] providers, String clientId, String clientSecret, String countryCode, AudioPlayerManager audioPlayerManager) {
-        this(clientId, clientSecret, countryCode, audioPlayerManager, new DefaultMirroringAudioTrackLookup(providers));
-    }
+	public SpotifySourceManager(String[] providers, String clientId, String clientSecret, String countryCode, AudioPlayerManager audioPlayerManager) {
+		this(clientId, clientSecret, countryCode, audioPlayerManager, new DefaultMirroringAudioTrackResolver(providers));
+	}
 
-    public SpotifySourceManager(String clientId, String clientSecret, String countryCode, AudioPlayerManager audioPlayerManager, MirroringAudioTrackLookup mirroringAudioTrackLookup) {
-        super(audioPlayerManager, mirroringAudioTrackLookup);
+	public SpotifySourceManager(String clientId, String clientSecret, String countryCode, AudioPlayerManager audioPlayerManager, MirroringAudioTrackResolver mirroringAudioTrackResolver) {
+		super(audioPlayerManager, mirroringAudioTrackResolver);
 
-        if (clientId == null || clientId.isEmpty()) {
-            throw new IllegalArgumentException("Spotify client id must be set");
-        }
-        this.clientId = clientId;
+		if (clientId == null || clientId.isEmpty()) {
+			throw new IllegalArgumentException("Spotify client id must be set");
+		}
+		this.clientId = clientId;
 
-        if (clientSecret == null || clientSecret.isEmpty()) {
-            throw new IllegalArgumentException("Spotify secret must be set");
-        }
-        this.clientSecret = clientSecret;
+		if (clientSecret == null || clientSecret.isEmpty()) {
+			throw new IllegalArgumentException("Spotify secret must be set");
+		}
+		this.clientSecret = clientSecret;
 
-        if (countryCode == null || countryCode.isEmpty()) {
-            countryCode = "US";
-        }
-        this.countryCode = countryCode;
-    }
+		if (countryCode == null || countryCode.isEmpty()) {
+			countryCode = "US";
+		}
+		this.countryCode = countryCode;
+	}
 
-    @Override
-    public String getSourceName() {
-        return "spotify";
-    }
+	@Override
+	public String getSourceName() {
+		return "spotify";
+	}
 
-    @Override
-    public AudioTrack decodeTrack(AudioTrackInfo trackInfo, DataInput input) throws IOException {
-        return new SpotifyAudioTrack(trackInfo,
-                DataFormatTools.readNullableText(input),
-                DataFormatTools.readNullableText(input),
-                this
-        );
-    }
+	@Override
+	public AudioTrack decodeTrack(AudioTrackInfo trackInfo, DataInput input) throws IOException {
+		return new SpotifyAudioTrack(trackInfo,
+				DataFormatTools.readNullableText(input),
+				DataFormatTools.readNullableText(input),
+				this
+		);
+	}
 
-    @Override
-    public AudioItem loadItem(AudioPlayerManager manager, AudioReference reference) {
-        try {
-            if (reference.identifier.startsWith(SEARCH_PREFIX)) {
-                return this.getSearch(reference.identifier.substring(SEARCH_PREFIX.length()).trim());
-            }
+	@Override
+	public AudioItem loadItem(AudioPlayerManager manager, AudioReference reference) {
+		try {
+			if (reference.identifier.startsWith(SEARCH_PREFIX)) {
+				return this.getSearch(reference.identifier.substring(SEARCH_PREFIX.length()).trim());
+			}
 
-            if (reference.identifier.startsWith(RECOMMENDATIONS_PREFIX)) {
-                return this.getRecommendations(reference.identifier.substring(RECOMMENDATIONS_PREFIX.length()).trim());
-            }
+			if (reference.identifier.startsWith(RECOMMENDATIONS_PREFIX)) {
+				return this.getRecommendations(reference.identifier.substring(RECOMMENDATIONS_PREFIX.length()).trim());
+			}
 
-            var matcher = URL_PATTERN.matcher(reference.identifier);
-            if (!matcher.find()) {
-                return null;
-            }
+			var matcher = URL_PATTERN.matcher(reference.identifier);
+			if (!matcher.find()) {
+				return null;
+			}
 
-            var id = matcher.group("identifier");
-            switch (matcher.group("type")) {
-                case "album":
-                    return this.getAlbum(id);
+			var id = matcher.group("identifier");
+			switch (matcher.group("type")) {
+				case "album":
+					return this.getAlbum(id);
 
-                case "track":
-                    return this.getTrack(id);
+				case "track":
+					return this.getTrack(id);
 
-                case "playlist":
-                    return this.getPlaylist(id);
+				case "playlist":
+					return this.getPlaylist(id);
 
-                case "artist":
-                    return this.getArtist(id);
-            }
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-        return null;
-    }
+				case "artist":
+					return this.getArtist(id);
+			}
+		} catch (IOException e) {
+			throw new RuntimeException(e);
+		}
+		return null;
+	}
 
-    public void requestToken() throws IOException {
-        var request = new HttpPost("https://accounts.spotify.com/api/token");
-        request.addHeader("Authorization", "Basic " + Base64.getEncoder().encodeToString((this.clientId + ":" + this.clientSecret).getBytes(StandardCharsets.UTF_8)));
-        request.setEntity(new UrlEncodedFormEntity(List.of(new BasicNameValuePair("grant_type", "client_credentials")), StandardCharsets.UTF_8));
+	public void requestToken() throws IOException {
+		var request = new HttpPost("https://accounts.spotify.com/api/token");
+		request.addHeader("Authorization", "Basic " + Base64.getEncoder().encodeToString((this.clientId + ":" + this.clientSecret).getBytes(StandardCharsets.UTF_8)));
+		request.setEntity(new UrlEncodedFormEntity(List.of(new BasicNameValuePair("grant_type", "client_credentials")), StandardCharsets.UTF_8));
 
-        var json = HttpClientTools.fetchResponseAsJson(this.httpInterfaceManager.getInterface(), request);
-        this.token = json.get("access_token").text();
-        this.tokenExpire = Instant.now().plusSeconds(json.get("expires_in").asLong(0));
-    }
+		var json = HttpClientTools.fetchResponseAsJson(this.httpInterfaceManager.getInterface(), request);
+		this.token = json.get("access_token").text();
+		this.tokenExpire = Instant.now().plusSeconds(json.get("expires_in").asLong(0));
+	}
 
-    public String getToken() throws IOException {
-        if (this.token == null || this.tokenExpire == null || this.tokenExpire.isBefore(Instant.now())) {
-            this.requestToken();
-        }
-        return this.token;
-    }
+	public String getToken() throws IOException {
+		if (this.token == null || this.tokenExpire == null || this.tokenExpire.isBefore(Instant.now())) {
+			this.requestToken();
+		}
+		return this.token;
+	}
 
-    public JsonBrowser getJson(String uri) throws IOException {
-        var request = new HttpGet(uri);
-        request.addHeader("Authorization", "Bearer " + this.getToken());
-        return HttpClientTools.fetchResponseAsJson(this.httpInterfaceManager.getInterface(), request);
-    }
+	public JsonBrowser getJson(String uri) throws IOException {
+		var request = new HttpGet(uri);
+		request.addHeader("Authorization", "Bearer " + this.getToken());
+		return HttpClientTools.fetchResponseAsJson(this.httpInterfaceManager.getInterface(), request);
+	}
 
-    public AudioItem getSearch(String query) throws IOException {
-        var json = this.getJson(API_BASE + "search?q=" + URLEncoder.encode(query, StandardCharsets.UTF_8) + "&type=track");
-        if (json == null || json.get("tracks").get("items").values().isEmpty()) {
-            return AudioReference.NO_TRACK;
-        }
+	public AudioItem getSearch(String query) throws IOException {
+		var json = this.getJson(API_BASE + "search?q=" + URLEncoder.encode(query, StandardCharsets.UTF_8) + "&type=track");
+		if (json == null || json.get("tracks").get("items").values().isEmpty()) {
+			return AudioReference.NO_TRACK;
+		}
 
-        return new BasicAudioPlaylist("Search results for: " + query, this.parseTrackItems(json.get("tracks")), null, true);
-    }
+		return new BasicAudioPlaylist("Search results for: " + query, this.parseTrackItems(json.get("tracks")), null, true);
+	}
 
-    public AudioItem getRecommendations(String query) throws IOException {
-        var json = this.getJson(API_BASE + "recommendations?" + query);
-        if (json == null || json.get("tracks").values().isEmpty()) {
-            return AudioReference.NO_TRACK;
-        }
+	public AudioItem getRecommendations(String query) throws IOException {
+		var json = this.getJson(API_BASE + "recommendations?" + query);
+		if (json == null || json.get("tracks").values().isEmpty()) {
+			return AudioReference.NO_TRACK;
+		}
 
-        return new BasicAudioPlaylist("Spotify Recommendations:", this.parseTracks(json), null, false);
-    }
+		return new BasicAudioPlaylist("Spotify Recommendations:", this.parseTracks(json), null, false);
+	}
 
-    public AudioItem getAlbum(String id) throws IOException {
-        var json = this.getJson(API_BASE + "albums/" + id);
-        if (json == null) {
-            return AudioReference.NO_TRACK;
-        }
+	public AudioItem getAlbum(String id) throws IOException {
+		var json = this.getJson(API_BASE + "albums/" + id);
+		if (json == null) {
+			return AudioReference.NO_TRACK;
+		}
 
-        var tracks = new ArrayList<AudioTrack>();
-        JsonBrowser page;
-        var offset = 0;
-        do {
-            page = this.getJson(API_BASE + "albums/" + id + "/tracks?limit=" + ALBUM_MAX_PAGE_ITEMS + "&offset=" + offset);
-            offset += ALBUM_MAX_PAGE_ITEMS;
+		var tracks = new ArrayList<AudioTrack>();
+		JsonBrowser page;
+		var offset = 0;
+		do {
+			page = this.getJson(API_BASE + "albums/" + id + "/tracks?limit=" + ALBUM_MAX_PAGE_ITEMS + "&offset=" + offset);
+			offset += ALBUM_MAX_PAGE_ITEMS;
 
-            var tracksPage = this.getJson(API_BASE + "tracks/?ids=" + page.get("items").values().stream().map(track -> track.get("id").text()).collect(Collectors.joining(",")));
+			var tracksPage = this.getJson(API_BASE + "tracks/?ids=" + page.get("items").values().stream().map(track -> track.get("id").text()).collect(Collectors.joining(",")));
 
-            tracks.addAll(this.parseTracks(tracksPage));
-        }
-        while (page.get("next").text() != null);
+			tracks.addAll(this.parseTracks(tracksPage));
+		}
+		while (page.get("next").text() != null);
 
-        if (tracks.isEmpty()) {
-            return AudioReference.NO_TRACK;
-        }
+		if (tracks.isEmpty()) {
+			return AudioReference.NO_TRACK;
+		}
 
-        return new BasicAudioPlaylist(json.get("name").text(), tracks, null, false);
+		return new BasicAudioPlaylist(json.get("name").text(), tracks, null, false);
 
-    }
+	}
 
-    public AudioItem getPlaylist(String id) throws IOException {
-        var json = this.getJson(API_BASE + "playlists/" + id);
-        if (json == null) {
-            return AudioReference.NO_TRACK;
-        }
+	public AudioItem getPlaylist(String id) throws IOException {
+		var json = this.getJson(API_BASE + "playlists/" + id);
+		if (json == null) {
+			return AudioReference.NO_TRACK;
+		}
 
-        var tracks = new ArrayList<AudioTrack>();
-        JsonBrowser page;
-        var offset = 0;
-        do {
-            page = this.getJson(API_BASE + "playlists/" + id + "/tracks?limit=" + PLAYLIST_MAX_PAGE_ITEMS + "&offset=" + offset);
-            offset += PLAYLIST_MAX_PAGE_ITEMS;
+		var tracks = new ArrayList<AudioTrack>();
+		JsonBrowser page;
+		var offset = 0;
+		do {
+			page = this.getJson(API_BASE + "playlists/" + id + "/tracks?limit=" + PLAYLIST_MAX_PAGE_ITEMS + "&offset=" + offset);
+			offset += PLAYLIST_MAX_PAGE_ITEMS;
 
-            for (var value : page.get("items").values()) {
-                var track = value.get("track");
-                if (track.isNull() || track.get("is_local").asBoolean(false)) {
-                    continue;
-                }
-                tracks.add(this.parseTrack(track));
-            }
+			for (var value : page.get("items").values()) {
+				var track = value.get("track");
+				if (track.isNull() || track.get("is_local").asBoolean(false)) {
+					continue;
+				}
+				tracks.add(this.parseTrack(track));
+			}
 
-        }
-        while (page.get("next").text() != null);
+		}
+		while (page.get("next").text() != null);
 
-        if (tracks.isEmpty()) {
-            return AudioReference.NO_TRACK;
-        }
+		if (tracks.isEmpty()) {
+			return AudioReference.NO_TRACK;
+		}
 
-        return new BasicAudioPlaylist(json.get("name").text(), tracks, null, false);
+		return new BasicAudioPlaylist(json.get("name").text(), tracks, null, false);
 
-    }
+	}
 
-    public AudioItem getArtist(String id) throws IOException {
-        var json = this.getJson(API_BASE + "artists/" + id + "/top-tracks?market=" + this.countryCode);
-        if (json == null || json.get("tracks").values().isEmpty()) {
-            return AudioReference.NO_TRACK;
-        }
-        return new BasicAudioPlaylist(json.get("tracks").index(0).get("artists").index(0).get("name").text() + "'s Top Tracks", this.parseTracks(json), null, false);
-    }
+	public AudioItem getArtist(String id) throws IOException {
+		var json = this.getJson(API_BASE + "artists/" + id + "/top-tracks?market=" + this.countryCode);
+		if (json == null || json.get("tracks").values().isEmpty()) {
+			return AudioReference.NO_TRACK;
+		}
+		return new BasicAudioPlaylist(json.get("tracks").index(0).get("artists").index(0).get("name").text() + "'s Top Tracks", this.parseTracks(json), null, false);
+	}
 
-    public AudioItem getTrack(String id) throws IOException {
-        var json = this.getJson(API_BASE + "tracks/" + id);
-        if (json == null) {
-            return AudioReference.NO_TRACK;
-        }
-        return parseTrack(json);
-    }
+	public AudioItem getTrack(String id) throws IOException {
+		var json = this.getJson(API_BASE + "tracks/" + id);
+		if (json == null) {
+			return AudioReference.NO_TRACK;
+		}
+		return parseTrack(json);
+	}
 
-    private List<AudioTrack> parseTracks(JsonBrowser json) {
-        var tracks = new ArrayList<AudioTrack>();
-        for (var value : json.get("tracks").values()) {
-            tracks.add(this.parseTrack(value));
-        }
-        return tracks;
-    }
+	private List<AudioTrack> parseTracks(JsonBrowser json) {
+		var tracks = new ArrayList<AudioTrack>();
+		for (var value : json.get("tracks").values()) {
+			tracks.add(this.parseTrack(value));
+		}
+		return tracks;
+	}
 
-    private List<AudioTrack> parseTrackItems(JsonBrowser json) {
-        var tracks = new ArrayList<AudioTrack>();
-        for (var value : json.get("items").values()) {
-            if (value.get("is_local").asBoolean(false)) {
-                continue;
-            }
-            tracks.add(this.parseTrack(value));
-        }
-        return tracks;
-    }
+	private List<AudioTrack> parseTrackItems(JsonBrowser json) {
+		var tracks = new ArrayList<AudioTrack>();
+		for (var value : json.get("items").values()) {
+			if (value.get("is_local").asBoolean(false)) {
+				continue;
+			}
+			tracks.add(this.parseTrack(value));
+		}
+		return tracks;
+	}
 
-    private AudioTrack parseTrack(JsonBrowser json) {
-        return new SpotifyAudioTrack(
-                new AudioTrackInfo(
-                        json.get("name").text(),
-                        json.get("artists").index(0).get("name").text(),
-                        json.get("duration_ms").asLong(0),
-                        json.get("id").text(),
-                        false,
-                        json.get("external_urls").get("spotify").text()
-                ),
-                json.get("external_ids").get("isrc").text(),
-                json.get("album").get("images").index(0).get("url").text(),
-                this
-        );
-    }
+	private AudioTrack parseTrack(JsonBrowser json) {
+		return new SpotifyAudioTrack(
+				new AudioTrackInfo(
+						json.get("name").text(),
+						json.get("artists").index(0).get("name").text(),
+						json.get("duration_ms").asLong(0),
+						json.get("id").text(),
+						false,
+						json.get("external_urls").get("spotify").text()
+				),
+				json.get("external_ids").get("isrc").text(),
+				json.get("album").get("images").index(0).get("url").text(),
+				this
+		);
+	}
 
-    @Override
-    public void shutdown() {
-        try {
-            this.httpInterfaceManager.close();
-        } catch (IOException e) {
-            log.error("Failed to close HTTP interface manager", e);
-        }
-    }
+	@Override
+	public void shutdown() {
+		try {
+			this.httpInterfaceManager.close();
+		} catch (IOException e) {
+			log.error("Failed to close HTTP interface manager", e);
+		}
+	}
 
-    @Override
-    public void configureRequests(Function<RequestConfig, RequestConfig> configurator) {
-        this.httpInterfaceManager.configureRequests(configurator);
-    }
+	@Override
+	public void configureRequests(Function<RequestConfig, RequestConfig> configurator) {
+		this.httpInterfaceManager.configureRequests(configurator);
+	}
 
-    @Override
-    public void configureBuilder(Consumer<HttpClientBuilder> configurator) {
-        this.httpInterfaceManager.configureBuilder(configurator);
-    }
+	@Override
+	public void configureBuilder(Consumer<HttpClientBuilder> configurator) {
+		this.httpInterfaceManager.configureBuilder(configurator);
+	}
 
 }
