+++ plugin/src/main/java/com/github/topisenpai/lavasrc/plugin/AppleMusicConfig.java
@@ -7,23 +7,23 @@ import org.springframework.stereotype.Component;
 @Component
 public class AppleMusicConfig {
 
-    private String countryCode = "us";
-    private String mediaAPIToken;
+	private String countryCode = "us";
+	private String mediaAPIToken;
 
-    public String getCountryCode() {
-        return this.countryCode;
-    }
+	public String getCountryCode() {
+		return this.countryCode;
+	}
 
-    public void setCountryCode(String countryCode) {
-        this.countryCode = countryCode;
-    }
+	public void setCountryCode(String countryCode) {
+		this.countryCode = countryCode;
+	}
 
-    public String getMediaAPIToken() {
-        return this.mediaAPIToken;
-    }
+	public String getMediaAPIToken() {
+		return this.mediaAPIToken;
+	}
 
-    public void setMediaAPIToken(String mediaAPIToken) {
-        this.mediaAPIToken = mediaAPIToken;
-    }
+	public void setMediaAPIToken(String mediaAPIToken) {
+		this.mediaAPIToken = mediaAPIToken;
+	}
 
 }
