+++ plugin/src/main/java/com/github/topisenpai/lavasrc/plugin/SpotifyConfig.java
@@ -7,32 +7,32 @@ import org.springframework.stereotype.Component;
 @Component
 public class SpotifyConfig {
 
-    private String clientId;
-    private String clientSecret;
-    private String countryCode;
+	private String clientId;
+	private String clientSecret;
+	private String countryCode;
 
-    public String getClientId() {
-        return this.clientId;
-    }
+	public String getClientId() {
+		return this.clientId;
+	}
 
-    public void setClientId(String clientId) {
-        this.clientId = clientId;
-    }
+	public void setClientId(String clientId) {
+		this.clientId = clientId;
+	}
 
-    public String getClientSecret() {
-        return this.clientSecret;
-    }
+	public String getClientSecret() {
+		return this.clientSecret;
+	}
 
-    public void setClientSecret(String clientSecret) {
-        this.clientSecret = clientSecret;
-    }
+	public void setClientSecret(String clientSecret) {
+		this.clientSecret = clientSecret;
+	}
 
-    public String getCountryCode() {
-        return this.countryCode;
-    }
+	public String getCountryCode() {
+		return this.countryCode;
+	}
 
-    public void setCountryCode(String countryCode) {
-        this.countryCode = countryCode;
-    }
+	public void setCountryCode(String countryCode) {
+		this.countryCode = countryCode;
+	}
 
 }
