+++ plugin/src/main/java/com/github/topisenpai/lavasrc/plugin/YandexMusicConfig.java
@@ -7,13 +7,13 @@ import org.springframework.stereotype.Component;
 @Component
 public class YandexMusicConfig {
 
-    private String accessToken;
+	private String accessToken;
 
-    public String getAccessToken() {
-        return this.accessToken;
-    }
+	public String getAccessToken() {
+		return this.accessToken;
+	}
 
-    public void setAccessToken(String accessToken) {
-        this.accessToken = accessToken;
-    }
+	public void setAccessToken(String accessToken) {
+		this.accessToken = accessToken;
+	}
 }
